E3 housekeeping

+++ E3/src/CheckersGridAccessor.java
@@ -1,9 +1,9 @@
 abstract class CheckersGridAccessor {
-    public final int maxOffset;
+    public final double maxOffset;
 
-    protected CheckersGridAccessor(int maxOffset) {
+    protected CheckersGridAccessor(double maxOffset) {
         this.maxOffset = maxOffset;
     }
 
-    int accessCheckersGrid(CheckersGrid checkersGrid, PlayerColor playerColor) { throw new IllegalStateException("Method not implemented"); }
+    double accessCheckersGrid(CheckersGrid checkersGrid, PlayerColor playerColor) { throw new IllegalStateException("Method not implemented"); }
 }
